qca: ipq: fix to correct wrong delay obtained from udelay function
If usec >= 224000 usec, then:
usec * GPT_FREQ_KHZ = 224000 * 19200 >= 0x100590000,
which is a 33 bits value

To fix this usec is typecasted to 64 bit
(unsigned long long)usec * GPT_FREQ_KHZ

(cherry picked from commit 72d8ec613fe67eb64569200627c8c2da80218563)

+++ arch/arm/cpu/armv7/qca/common/timer.c
@@ -153,7 +153,7 @@ void __udelay(unsigned long usec)
 	if (usec == 0)
 		return;
 
-	val = (usec * GPT_FREQ_KHZ) / 1000;
+	val = ((unsigned long long)usec * GPT_FREQ_KHZ) / 1000;
 	if (val == 0 )
 		val = 1; /* Wait for atleast 1 tick */
